// api/src/dataschema/Naruto/INarutoPersistence.ts
export interface INarutoPersistence {
    domainId: string;
    name: string;
    gender: string;
    affiliation: string[];
    jutsu_type: string[];
    kekkei_genkai: string[];
    nature_type: string[];
    attribute: string[];
    debut_arc: string;
    vital_state: string;
    occupation: string[];
}